requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0,<3.0
pydantic>=2.0.0,<3.0
python-dotenv>=1.0.0
pymysql>=1.1.0
openai>=1.40.0